# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GAMEA Social Monitor",
    "description": "API Institucional de Monitoreo y Analítica de Interacciones en Redes Sociales del Gobierno Autónomo Municipal de El Alto (GAMEA).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/interactions/": {
      "post": {
        "tags": [
          "Interactions & Evidence"
        ],
        "summary": "Registrar o ingerir interacción de forma idempotente (REQ-INT-001)",
        "operationId": "ingest_interaction_api_v1_interactions__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Interactions & Evidence"
        ],
        "summary": "Listar interacciones observadas con filtros",
        "operationId": "list_interactions_api_v1_interactions__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publication Id"
            }
          },
          {
            "name": "platform_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform Id"
            }
          },
          {
            "name": "interaction_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interaction Type"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página (1-indexado)",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página (1-indexado)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Registros por página (máximo 1000)",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Registros por página (máximo 1000)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_InteractionResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interactions/{interaction_id}": {
      "get": {
        "tags": [
          "Interactions & Evidence"
        ],
        "summary": "Obtener detalle de una interacción",
        "operationId": "get_interaction_api_v1_interactions__interaction_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Interaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/interactions/users/list": {
      "get": {
        "tags": [
          "Interactions & Evidence"
        ],
        "summary": "Listar usuarios únicos que interactuaron en redes sociales (comentarios y reacciones)",
        "description": "Retorna la lista de usuarios únicos que realizaron comentarios o reacciones,\ncon conteo desglosado por tipo de interacción y plataformas utilizadas.",
        "operationId": "list_interaction_users_api_v1_interactions_users_list_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por publicación específica",
              "title": "Publication Id"
            },
            "description": "Filtrar por publicación específica"
          },
          {
            "name": "platform_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por plataforma",
              "title": "Platform Id"
            },
            "description": "Filtrar por plataforma"
          },
          {
            "name": "interaction_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por tipo: COMMENT, LIKE, SHARE, REPLY",
              "title": "Interaction Type"
            },
            "description": "Filtrar por tipo: COMMENT, LIKE, SHARE, REPLY"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InteractionUserItem"
                  },
                  "title": "Response List Interaction Users Api V1 Interactions Users List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InteractionCreate": {
        "properties": {
          "publication_id": {
            "type": "string",
            "format": "uuid",
            "title": "Publication Id"
          },
          "platform_id": {
            "type": "string",
            "format": "uuid",
            "title": "Platform Id"
          },
          "interaction_type": {
            "type": "string",
            "title": "Interaction Type",
            "description": "COMMENT, LIKE, SHARE, REPLY"
          },
          "external_interaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Interaction Id",
            "description": "ID único externo del comentario o reacción"
          },
          "external_post_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Post Id"
          },
          "external_author_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Author Id",
            "description": "ID social del autor (Graph API / TikTok API)"
          },
          "external_author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Author Name"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "reaction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reaction Type"
          },
          "external_created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Created At"
          },
          "capture_method": {
            "type": "string",
            "title": "Capture Method",
            "default": "POLLING"
          },
          "api_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Version",
            "default": "v26.0"
          },
          "data_origin_type": {
            "type": "string",
            "title": "Data Origin Type",
            "default": "EMPLOYEE_INTERACTION_OFFICIAL"
          },
          "source_platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Platform"
          },
          "source_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Account Id"
          },
          "raw_payload": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Payload",
            "description": "Payload JSON crudo para generación de evidencia criptográfica"
          }
        },
        "type": "object",
        "required": [
          "publication_id",
          "platform_id",
          "interaction_type"
        ],
        "title": "InteractionCreate"
      },
      "InteractionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "publication_id": {
            "type": "string",
            "format": "uuid",
            "title": "Publication Id"
          },
          "platform_id": {
            "type": "string",
            "format": "uuid",
            "title": "Platform Id"
          },
          "interaction_type": {
            "type": "string",
            "title": "Interaction Type"
          },
          "external_interaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Interaction Id"
          },
          "external_post_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Post Id"
          },
          "external_author_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Author Id"
          },
          "external_author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Author Name"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "reaction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reaction Type"
          },
          "external_created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Created At"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "capture_method": {
            "type": "string",
            "title": "Capture Method"
          },
          "api_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Version"
          },
          "data_origin_type": {
            "type": "string",
            "title": "Data Origin Type"
          },
          "source_platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Platform"
          },
          "source_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Account Id"
          },
          "raw_payload_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Payload Ref"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "publication_id",
          "platform_id",
          "interaction_type",
          "captured_at",
          "capture_method",
          "data_origin_type",
          "created_at"
        ],
        "title": "InteractionResponse"
      },
      "InteractionUserItem": {
        "properties": {
          "external_author_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Author Id"
          },
          "external_author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Author Name"
          },
          "total_interactions": {
            "type": "integer",
            "title": "Total Interactions",
            "default": 0
          },
          "comments_count": {
            "type": "integer",
            "title": "Comments Count",
            "default": 0
          },
          "reactions_count": {
            "type": "integer",
            "title": "Reactions Count",
            "default": 0
          },
          "shares_count": {
            "type": "integer",
            "title": "Shares Count",
            "default": 0
          },
          "platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Platforms",
            "default": []
          },
          "first_interaction_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Interaction At"
          },
          "last_interaction_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Interaction At"
          }
        },
        "type": "object",
        "title": "InteractionUserItem",
        "description": "Usuario único que interactuó en redes sociales (comentario o reacción)."
      },
...
      "PageResponse_InteractionResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InteractionResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Lista de elementos de la página actual"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total general de elementos"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "Número de página actual"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size",
            "description": "Tamaño de la página"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "description": "Total de páginas disponibles"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PageResponse[InteractionResponse]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}